Fix power calculation when accounting for gaps
Was calculating e.g. 2 second power with 3 seconds' worth of power due to
the eager calculation when walking left edge up. Instead, do this eager
calculation only if the next reading will overshoot the target duration, so
we know to check so that the accumulated power total isn't just ignored.